# Ensure fixed-opaque types (arrays) implement EncodeTo() by pointer