@leighmcculloch 2021-11-10
The diff you posted was so useful in understanding the change that I think we should add a CI run that runs against stellar/go and prints out the diff. I created https://github.com/stellar/xdrgen/issues/69.